+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLListSelectionField.cs
@@ -95,7 +95,7 @@ namespace EntityGraphQL.Compiler
                 {
                     // To support a common use case where we are coming from a service result to another service field where the 
                     // service is the Query Context. Which we are assuming is likely an EF context and we don't need the null check
-                    // Use ExecutionOptions.ExecuteServiceFieldsSeparately = false to disable this behaviour
+                    // Use ExecutionOptions.ExecuteServiceFieldsSeparately = false to disable this behavior
                     var nullCheck = Field!.Services.Any(s => s.Type != Field.Schema.QueryContextType);
                     (resultExpression, PossibleNextContextTypes) = ExpressionUtil.MakeSelectWithDynamicType(this, nextFieldContext!, listContext, selectionFields, nullCheck, withoutServiceFields);
                 }