+++ src/EntityGraphQL/Schema/IField.cs
@@ -69,6 +69,7 @@ namespace EntityGraphQL.Schema
         /// </summary>
         List<ParameterExpression> Services { get; }
         IReadOnlyCollection<Action<ArgumentValidatorContext>> Validators { get; }
+        [Obsolete("Avoid using this method, it creates issues if the field's type is used on multiple fields with different arguments. It will be removed in future versions. See updated OffsetPagingExtension for a better way using ProcessArguments")]
         IField? UseArgumentsFromField { get; }
 
         /// <summary>
@@ -81,7 +82,6 @@ namespace EntityGraphQL.Schema
         /// <returns></returns>
         (Expression? expression, ParameterExpression? argumentParam) GetExpression(Expression fieldExpression, Expression? fieldContext, IGraphQLNode? parentNode, ParameterExpression? schemaContext, CompileContext? compileContext, IReadOnlyDictionary<string, object> args, ParameterExpression? docParam, object? docVariables, IEnumerable<GraphQLDirective> directives, bool contextChanged, ParameterReplacer replacer);
         Expression? ResolveExpression { get; }
-
         IField UpdateExpression(Expression expression);
 
         IField AddExtension(IFieldExtension extension);
@@ -91,6 +91,7 @@ namespace EntityGraphQL.Schema
         /// <param name="args"></param>
         void AddArguments(object args);
         IField Returns(GqlTypeInfo gqlTypeInfo);
+        [Obsolete("Avoid using this method, it creates issues if the field's type is used on multiple fields with different arguments. It will be removed in future versions. See updated OffsetPagingExtension for a better way using ProcessArguments")]
         void UseArgumentsFrom(IField field);
         IField AddValidator<TValidator>() where TValidator : IArgumentValidator;
         IField AddValidator(Action<ArgumentValidatorContext> callback);