+++ CHANGELOG.md
@@ -1,8 +1,36 @@
+# 5.3.0
+
+## Changes
+
+- `IField.UseArgumentsFrom` & `IField.UseArgumentsFromField` is now Obsolete. Using it (typically in a field extension) creates issues. See #358 fix below
+- `IFieldExtension.ProcessArguments` has been added to allow field extensions to handle arguments and add them to the compile context. This interface may merge with the existing `IFieldExtension.GetExpression` in version 6.0. This was introduced to fix #358 and avoid breaking changes
+
+## Fixes
+
+- #358 - `OffsetPaging` and `ConnectionPaging` field extensions can now be used on multiple fields that return the same type. e.g.
+
+```
+// both fields return a list of people but have different expressions
+schema.Query().AddField("peopleOver", new
+  {
+      over = ArgumentHelper.Required<int>()
+  },
+  (ctx, args) => ctx.People.Where(p => p.Height > args.over).OrderBy(p => p.Id), "Return list of people with paging metadata")
+  .UseOffsetPaging();
+schema.Query().AddField("peopleUnder", new
+  {
+      under = ArgumentHelper.Required<int>()
+  },
+  (ctx, args) => ctx.People.Where(p => p.Height < args.under).OrderBy(p => p.Id), "Return list of people with paging metadata")
+  .UseOffsetPaging();
+```
+
 # 5.2.1
 
 ## Fixes
 
 - #355 - make error variable values more distinctive
+- EntityGraphQL is now more specific in which methods it tries to optimize the pattern of a list to single. E.g. If a field has an expression like `ctx => ctx.People.FirstOrDefault(filter)` EntityGraphQL treats that as a ListToSingle field and will build a full expression similar to `ctx => ctx.People.Where(filter).Select(x => new {...}).FirstOrDefault()` to avoid over fetching when using EntityFramework. Previously EntityGraphQL would try to break up and rebuild expressions.methods it didn't know how to. It now correctly only tries this optimization on `First`, `FirstOrDefault`, `Single`, `SingleOrDefault`, `Last` & `LastOrDefault` from `Queryable` & `Enumerable`
 
 # 5.2.0